+++ lib/flutter_multi_formatter.dart
@@ -26,14 +26,15 @@ THE SOFTWARE.
 
 library flutter_multi_formatter;
 
-export 'formatters/formatter_utils.dart';
-export 'formatters/masked_input_formatter.dart';
-export 'formatters/phone_input_enums.dart';
-export 'formatters/phone_input_formatter.dart';
-export 'formatters/money_input_formatter.dart';
-export 'formatters/money_input_enums.dart';
-export 'formatters/formatter_extension_methods.dart';
 export 'formatters/credit_card_cvc_input_formatter.dart';
 export 'formatters/credit_card_expiration_input_formatter.dart';
 export 'formatters/credit_card_number_input_formatter.dart';
+export 'formatters/formatter_extension_methods.dart';
+export 'formatters/formatter_utils.dart';
+export 'formatters/masked_input_formatter.dart';
+export 'formatters/money_input_enums.dart';
+export 'formatters/money_input_formatter.dart';
+export 'formatters/phone_input_enums.dart';
+export 'formatters/phone_input_formatter.dart';
+export 'formatters/restricting_formatter.dart';
 export 'utils/unfocuser.dart';